chore: restructured api.js folder structure
